backend/pyproject.toml
[tool.poetry]
name = "college-matcher-backend"
version = "0.1.0"
description = "FastAPI backend for College Matcher"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
snowflake-connector-python = "^3.10.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


